starting timestapms

+++ app/config/db/knex/migrations/20160809145730_init.js
@@ -14,7 +14,7 @@ return Promise.all([
     table.string('feed_content', 1000)
     table.string('feed_title')
     table.string('feed_identifier')
-    table.integer('created_at')    
+    table.bigint('created_at')    
   }),
   knex.schema.createTable('feedz_posts', function(table){
     table.increments('id').primary()
@@ -24,7 +24,7 @@ return Promise.all([
     table.string('post_title')
     table.string('post_image_url')
     table.string('post_external_url')
-    table.integer('created_at')
+    table.bigint('created_at')
   }),
   knex.schema.createTable('feedz_comments', function(table){
     table.increments('id').primary()
@@ -32,13 +32,13 @@ return Promise.all([
     table.integer('feedz_post_id').references("feedz_posts.id")
     table.string('comment_title')
     table.string('comment_content')
-    table.integer('created_at')
+    table.bigint('created_at')
   }),
   knex.schema.createTable('feedz_favorite_posts', function(table){
     table.increments('id').primary()
     table.integer('feedz_user_id').references("feedz_users.id")
     table.integer('feedz_post_id').references("feedz_posts.id")
-    table.integer('created_at')
+    table.bigint('created_at')
   }),
   knex.schema.createTable('feedz_karma', function(table){
     table.increments('id').primary()
@@ -46,7 +46,7 @@ return Promise.all([
     table.integer('feedz_post_id').references("feedz_posts.id")
     table.boolean('upvote')
     table.boolean('downvote')
-    table.integer('created_at')
+    table.bigint('created_at')
   }),
 ])}
 exports.down = function(knex, Promise) {}
